requirements.txt
pandas==2.2.3
numpy==2.2.5
psycopg2-binary==2.9.10
argparse==1.1
rapidfuzz==3.13.0
